refactor: move command functions into module
