+++ src/lib.rs
@@ -1,0 +1,5 @@
+pub mod commands;
+
+pub use commands::{
+    add_memo, archive_category, edit_memo, list_categories, list_memos, remove_memos,
+};
